requirements.txt
flask==2.3.3
openai-whisper==20231117
requests==2.31.0
torch>=2.2.0
numpy>=1.24.3 